https://issues.apache.org/jira/browse/AMQ-3607 - added test case and use of connectionFactory.setOptimizeAcknowledgeTimeOut, the test is breaking becauee of message discarding when the acks are not immediate, made some small mods to make the test work and fixed up the setter for connectionFactory.setOptimizeAcknowledgeTimeOut
git-svn-id: https://svn.apache.org/repos/asf/activemq/trunk@1208048 13f79535-47bb-0310-9956-ffa450edef68

+++ activemq-core/src/main/java/org/apache/activemq/ActiveMQConnection.java
@@ -1635,7 +1635,7 @@ public class ActiveMQConnection implements Connection, TopicConnection, QueueCon
      * The max time in milliseconds between optimized ack batches
      * @param optimizeAcknowledgeTimeOut
      */
-    public void setOptimizeAcknowledgeTimeOut(int optimizeAcknowledgeTimeOut) {
+    public void setOptimizeAcknowledgeTimeOut(long optimizeAcknowledgeTimeOut) {
         this.optimizeAcknowledgeTimeOut =  optimizeAcknowledgeTimeOut;
     }
